src/fastapi_app/requirements.txt
fastapi[all]
uvicorn
pillow
torchvision